Change Connection constructor to take sockets
The connection needs only request and response sockets as dependency
for proper working. For the most simple case Connection working
with one instance of mongrel2 can be created using Connection.for
method instead of Connection.new :

  Connection.for(sender_id, request_addr, response_addr, context)

Standard method using

  Connection.new(request_socket, response_socket)

is used for more general usecase. For example if someone want the
handler to work with two mongrel2 instances the sockets can be
created "by hand" and connected to both mongrels.

That way we have standard constructor for general usage and
delegated for the most common case that takes identical arguments
as initialize did before this refactoring.