Introduce ConnectionFactory
In the future we would like to use multiple threads. Every thread
will obtain requests from separate `ZMQ` connection as it is
recommended in the tutorial. Because of that we will have to
create the sockets for every thread. For that reason `Handler` takes
now in constructor a factory capable of creating new `Connection`
instead of the `Connection` itself.

`Connection#connection` method is introduced so that `Connection` has
interface compatible with `ConnectionFactory` and can be used
instead returning always `self` whenever `connection` is needed. That
should work fine for simplest testing case or in single-threaded
environment.